Reduced size of release.js with util files
