+++ package.json
@@ -59,6 +59,7 @@
     "@babel/preset-flow": "^7.0.0",
     "@babel/register": "^7.0.0",
     "@onfido/surge": "^0.20.2",
+    "app-root-path": "^2.2.1",
     "autoprefixer": "^7.1.1",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.0.1",
