+++ release/RELEASE_GUIDELINES.md
@@ -7,13 +7,13 @@ We follow semver for versioning. Given a version number MAJOR.MINOR.PATCH, incre
 
 # Prerequisite
 In order to perform the release you need to:
-1. Download the Web SDK release configuration file from Confluence and store it in this repository as `releaseConfig.js`
-  - Edit `releaseConfig.js` if necessary
+1. Download the Web SDK release configuration secret file and store it as `release/releaseSecret.js`
+  - Edit secret file where necessary
 2. Make sure your laptop is set up to be able to login to S3
 3. Make sure that your git configuration allows you to use git commands without inputting password or passphrase
 
 # Usage
-Run `VERSION=x.x.x node release.js`
+Run `VERSION=x.x.x node release/release.js`
 
 # After the release
 1. Merge release/<release_version> PR to master
