+++ release/release.js
@@ -1,137 +1,18 @@
 #!/usr/bin/env node
-const config = require('./releaseConfig')
-const yn = require('yn')
-const ora = require('ora')
-const readline = require('readline')
-const fs = require('fs')
-const { exec, spawn } = require('child_process')
-const util = require('util')
-const promiseExec = util.promisify(exec)
 const chalk = require('chalk')
 
+const config = require('./releaseConfig')
+const terminalUI = require('./utils/terminalUI')
+const io = require('./utils/IO')
+const file = require('./utils/file')
+const processes = require('./utils/processes')
+
+const { welcomeMessage, stepTitle } = terminalUI
+const { question, getNumberInput, proceedYesNo } = io
+const { spawnAssumeOkay, execAssumeOkay, execWithErrorHandling, exitRelease} = processes
+const { replaceInFile, readInFile } = file
+
 const { VERSION } = process.env
-
-let safeToClearWorkspace = false
-let rcNumber = NaN
-let versionRC = null
-let isFirstReleaseIteration = false
-
-//////////////////////////////
-// Release Helper functions //
-/////////////////////////////
-
-const stepTitle = message => {
-  console.log()
-  console.log(chalk.magenta('~'.repeat(message.length + 4)))
-  console.log(chalk.magenta(`| ${message} |`))
-  console.log(chalk.magenta('~'.repeat(message.length + 4)))
-  console.log()
-}
-
-const question = query => {
-  const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout
-  });
-
-  console.log()
-  return new Promise(resolve => rl.question(`\n${query} (y/n) `, answer => {
-    const answerAsBoolean = yn(answer) || false
-    resolve(answerAsBoolean)
-    rl.close()
-  }))
-}
-
-const getNumberInput = query => {
-  const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout
-  });
-
-  return new Promise(resolve => rl.question(`\n${query}`, answer => {
-    const answerAsNumber = parseInt(answer, 10)
-    rl.close()
-    resolve(answerAsNumber)
-  }))
-  .then((answerAsNumber) => {
-    if (!answerAsNumber) {
-      console.log(`❌ That was not a valid integer. Please type a valid integer 🙄 \n`)
-      return getNumberInput(query)
-    }
-    return answerAsNumber
-  })
-}
-
-const proceedYesNo = async query => {
-  const ok = await question(query || 'Is this correct?')
-  if (ok) {
-    console.log('✅ Great!\n')
-  } else {
-    console.error('❌ Things were not correct. I don\'t know how to automate this case 🤖😞')
-    exitRelease()
-  }
-}
-
-const spawnAssumeOkay = async (cmd, cmdArgs, verbose) => {
-  const spinner = ora([cmd].concat(cmdArgs).join(' '))
-  if (!verbose) {
-    spinner.start()
-  }
-
-  let exitInProcess = false
-  const handleExit = error => {
-    if (exitInProcess) return
-    exitInProcess = true
-
-    spinner.fail()
-    somethingWentWrong(cmd)
-    if (error) {
-      console.error(`${error}\n`)
-    }
-    exitRelease()
-  }
-
-  await new Promise(resolve => {
-    const handle = spawn(cmd, cmdArgs, { cwd: '.' })
-    if (verbose) {
-      handle.stdout.pipe(process.stdout);
-    }
-    handle.stderr.pipe(process.stderr);
-
-    const onClose = code => {
-      if (code === 0) {
-        spinner.succeed()
-        resolve()
-      } else {
-        handleExit()
-      }
-    }
-    handle.on('close', onClose)
-    handle.on('exit', onClose)
-
-    handle.on('error', handleExit)
-  })
-}
-
-const execAssumeOkay = async cmd => {
-  const spinner = ora(cmd).start()
-
-  try {
-    const ret = await promiseExec(cmd)
-    spinner.stop()
-    return ret
-  } catch (error) {
-    spinner.stop()
-    somethingWentWrong(cmd)
-    console.error(error)
-    exitRelease()
-  }
-}
-
-const somethingWentWrong = (cmd) => {
-  console.error('❌ Oops. Something went wrong with that last command! 🤖😞 \n')
-  console.error(`❌ The command was: ${chalk.magenta(cmd)} \n`)
-}
 
 const checkWorkspaceIsClean = async () => {
   const { stdout: workspaceIsUnclean } = await execAssumeOkay('git diff-index --quiet HEAD -- || echo "not clean"')
@@ -139,79 +20,14 @@ const checkWorkspaceIsClean = async () => {
     console.error('❌ Your git workspace must be clean before starting a release 🤖😞')
     exitRelease()
   }
-  safeToClearWorkspace = true
-}
-
-const replaceInFile = (file, regex, replaceFunc) => {
-  fs.readFile(file, 'utf8', (err, data) => {
-    if (err) {
-      console.error('❌ Something went wrong trying to load the file!')
-      console.error(err)
-      exitRelease()
-    }
-
-    const result = data.replace(regex, replaceFunc)
-
-    fs.writeFile(file, result, 'utf8',  (err) => {
-       if (err) {
-         console.error('❌ Something went wrong trying to write to the file!')
-         console.error(err)
-         exitRelease()
-       }
-    })
-  })
-}
-
-const readInFile = async (file, regex, callback) => {
-  fs.readFile(file, 'utf8', (err, data) => {
-    if (err) {
-      console.error('❌ Something went wrong trying to load the file!')
-      console.error(err)
-      exitRelease()
-    }
-    const result = data.match(regex)
-    callback(result)
-  })
-}
-
-const execWithErrorHandling = async (cmd, callback) => {
-  const spinner = ora(cmd).start()
-
-  try {
-    const ret = await promiseExec(cmd)
-    spinner.stop()
-    return ret
-  } catch (error) {
-    spinner.stop()
-    callback()
-  }
-}
-
-const exitRelease = async () => {
-  if (safeToClearWorkspace) {
-    console.log('Clearing any workspace changes introduced by the release script...')
-    await promiseExec('git checkout -- \'*\'')
-  }
-  process.exit(1)
-
-  // make sure this step never resolves, so `process.exit` calls before this
-  // function is "finished" (otherwise later steps can still get called)
-  await new Promise()
-}
-
-///////////////////
-// Release Steps //
-//////////////////
-
-const welcomeMessage = () => {
-  console.log('Beep boop. Release Bot at your service. Let\'s release the SDK 🤖👋')
+  config.write('safeToClearWorkspace', true)
 }
 
 const checkRequiredParams = () => {
   const required = ['VERSION']
   const missingEnvKeys = required.filter(reqEnv => !process.env[reqEnv])
   if (missingEnvKeys.length) {
-    console.error(`These are required environment variables! ${missingEnvKeys.join(', ')}`)
+    console.error(`❌ Some required environment variables are missing! ${missingEnvKeys.join(', ')}`)
     exitRelease()
   }
 }
@@ -220,11 +36,12 @@ const confirmReleaseVersion = async () => {
   stepTitle('Release Type & Version')
   const isReleaseCandidate = await question('Is this a Release Candidate?')
   if (isReleaseCandidate) {
-    rcNumber = await getNumberInput('What is the Release Candidate number? ')
-    versionRC = `${VERSION}-rc.${rcNumber}`
-    isFirstReleaseIteration = parseInt(rcNumber, 10) === 1
+    const rcNumber = await getNumberInput('What is the Release Candidate number? ')
+    config.write('versionRC', `${VERSION}-rc.${rcNumber}`)
+    const isFirstReleaseIteration = parseInt(rcNumber, 10) === 1
+    config.write('isFirstReleaseIteration', isFirstReleaseIteration)
     console.log(`This is a ${chalk.bold.yellow('RELEASE CANDIDATE')}.`)
-    console.log(`Version Candidate: "${chalk.bold.yellow(versionRC)}"`)
+    console.log(`Version Candidate: "${chalk.bold.yellow(config.data.versionRC)}"`)
     console.log(`Version (to eventually be part of): "${chalk.bold.yellow(VERSION)}"`)
   } else {
     console.log(`This is a ${chalk.bold.green('FULL')} release.`)
@@ -256,7 +73,7 @@ const letsGetStarted = () => {
 
 const checkoutAndPullLatestCode = async () => {
   stepTitle('👀 Checking out the latest branch...')
-  const branchToCheckout = isFirstReleaseIteration ? 'development' : `release/${VERSION}`
+  const branchToCheckout = config.data.isFirstReleaseIteration ? 'development' : `release/${VERSION}`
   console.log(`Great, checking out ${chalk.magenta(branchToCheckout)}`)
   await spawnAssumeOkay('git', ['checkout', branchToCheckout])
   await spawnAssumeOkay('git', ['pull', 'origin', branchToCheckout])
@@ -276,10 +93,10 @@ const incrementBase32Version = async () => {
   // The base32 should only be upfated once per release.
   // So we do it only for the first release candidate,
   // all the following iteration will use the same rc
-  if (isFirstReleaseIteration) {
+  if (config.data.isFirstReleaseIteration) {
     replaceInFile(
-      './webpack.config.babel.js',
-      /'BASE_32_VERSION'\s+: '([A-Z]+)'/,
+      '../../webpack.config.babel.js',
+      /'BASE_32_VERSION'(|\s): '([A-Z]+)'/,
       (_, groupMatch) => `'BASE_32_VERSION': '${bumpBase32(groupMatch)}'`
     )
   }
@@ -290,9 +107,9 @@ const incrementPackageJsonVersion = async () => {
   stepTitle('⬆️ Setting the package.json version...')
 
   replaceInFile(
-    './package.json',
+    '../../package.json',
     /"version": ".*"/,
-    () => `"version": "${versionRC || VERSION}"`
+    () => `"version": "${config.data.versionRC || VERSION}"`
   )
 
   console.log('✅ Success!')
@@ -300,10 +117,10 @@ const incrementPackageJsonVersion = async () => {
 
 const incrementVersionInJSFiddle = async () => {
   stepTitle('⬆️ Increment version in JSFiddle...')
-  const version = versionRC ? versionRC : VERSION
+  const version = config.data.versionRC ? config.data.versionRC : VERSION
 
   replaceInFile(
-    './demo/fiddle/demo.details',
+    '../../demo/fiddle/demo.details',
     /- https:\/\/assets\.onfido\.com\/web-sdk-releases\/.*\/onfido\.min\.js\n\s{3}- https:\/\/assets\.onfido\.com\/web-sdk-releases\/.*\/style\.css/,
     () => `- https://assets.onfido.com/web-sdk-releases/${version}/onfido.min.js\n${' '.repeat(3)}- https://assets.onfido.com/web-sdk-releases/${version}/style.css`
   )
@@ -362,7 +179,7 @@ const checkoutExistingReleaseBranch = async () => {
 }
 
 const checkoutOrCreateBranch = async () => {
-  if (isFirstReleaseIteration) {
+  if (config.data.isFirstReleaseIteration) {
     await createReleaseBranch()
   }
   else {
@@ -373,7 +190,7 @@ const checkoutOrCreateBranch = async () => {
 const makeReleaseCommit = async () => {
   stepTitle('💾 Making commit release')
 
-  const commitMessage = `Bump version to ${versionRC || VERSION}`
+  const commitMessage = `Bump version to ${config.data.versionRC || VERSION}`
   console.log(`Creating the commit message: "${commitMessage}"`)
 
   const isVerboseCmd = true
@@ -388,8 +205,8 @@ const makeReleaseCommit = async () => {
 const loginToS3 = async () => {
   stepTitle('🔐 Sign in to 1Password and S3')
   console.log('On another shell, please run the following commands:')
-  console.log(`${chalk.bold.yellow(config.OP_LOGIN_CMD)}`)
-  console.log(`${chalk.bold.yellow(config.S3_LOGIN_CMD)}`)
+  console.log(`${chalk.bold.yellow(config.data.OP_LOGIN_CMD)}`)
+  console.log(`${chalk.bold.yellow(config.data.S3_LOGIN_CMD)}`)
   await proceedYesNo('Have all of these commands succeeded?\n')
 }
 
@@ -397,12 +214,12 @@ const uploadToS3 = async () => {
   stepTitle('📤 Upload to S3')
   console.log('On another shell, please run the following commands:')
   // HACK: I wasn't able to access/store the base32 in any other way, therefore I had to use this hack
-  await readInFile('./webpack.config.babel.js',
-    /'BASE_32_VERSION': '([A-Z]+)'/,
+  await readInFile('../../webpack.config.babel.js',
+    /'BASE_32_VERSION'(|\s): '([A-Z]+)'/,
     (matchGroup) => {
-      console.log(`${chalk.bold.yellow(`${config.UPLOAD_CMD} ${config.S3_BUCKET}${config.BASE_32_FOLDER_PATH}/${matchGroup[1]}/`)}`)
-      const versionPath = versionRC ? versionRC : VERSION
-      console.log(`${chalk.bold.yellow(`${config.UPLOAD_CMD} ${config.S3_BUCKET}${config.RELEASES_FOLDER_PATH}/${versionPath}/`)}`)
+      console.log(`${chalk.bold.yellow(`${config.data.UPLOAD_CMD} ${config.data.S3_BUCKET}${config.data.BASE_32_FOLDER_PATH}/${matchGroup[1]}/`)}`)
+      const versionPath = config.data.versionRC ? config.data.versionRC : VERSION
+      console.log(`${chalk.bold.yellow(`${config.data.UPLOAD_CMD} ${config.S3_BUCKET}${config.data.RELEASES_FOLDER_PATH}/${versionPath}/`)}`)
     }
   )
   await new Promise(resolve => setTimeout(resolve, 1000))
@@ -413,8 +230,8 @@ const didS3uploadSucceed = async () => {
 }
 
 const publishTag = async () => {
-  if (versionRC) {
-    stepTitle(`🕑 Creating next tag for release candidate ${versionRC}`)
+  if (config.data.versionRC) {
+    stepTitle(`🕑 Creating next tag for release candidate ${config.data.versionRC}`)
     await spawnAssumeOkay('npm', ['publish', '--tag', 'next'])
     console.log('Done. Now make sure that the latest tag has not changed, only the next one:')
 
@@ -461,8 +278,8 @@ const publishOnNpm = async () => {
 
 const upgradeDemoAppToTag = async () => {
   stepTitle('🕑 Upgrading demo app...')
-  const versionToInstall = versionRC ? versionRC : VERSION
-  await spawnAssumeOkay('cd', [config.SAMPLE_APP_PATH])
+  const versionToInstall = config.data.versionRC ? config.data.versionRC : VERSION
+  await spawnAssumeOkay('cd', [config.data.SAMPLE_APP_PATH])
 
   const isVerboseCmd = true
   await spawnAssumeOkay('pwd',[], isVerboseCmd)
@@ -484,8 +301,9 @@ const releaseComplete = () => {
 }
 
 const main = async () => {
+  console.log(process.env.HOME)
   welcomeMessage()
-  safeToClearWorkspace = await checkWorkspaceIsClean()
+  await checkWorkspaceIsClean()
   checkRequiredParams()
   await confirmReleaseVersion()
   await confirmDocumentationCorrect()
@@ -505,7 +323,7 @@ const main = async () => {
   await didS3uploadSucceed()
   await npmLogin()
   await publishTag()
-  if (versionRC) {
+  if (config.data.versionRC) {
     await upgradeDemoAppToTag()
     regressionTesting()
   }
