+++ release/releaseConfig.js
@@ -1,0 +1,13 @@
+const secret = require('./releaseSecret')
+
+const Config = module.exports = {
+  data: {
+    ...secret,
+    safeToClearWorkspace: false,
+    versionRC: null,
+    isFirstReleaseIteration: false
+  },
+  write(key, value) {
+    Config.data[key] = value;
+  }
+}
